@@ -14,24 +14,22 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~7.1.0",
-    "@angular/common": "~7.1.0",
-    "@angular/compiler": "~7.1.0",
-    "@angular/core": "~7.1.0",
-    "@angular/forms": "~7.1.0",
-    "@angular/platform-browser": "~7.1.0",
-    "@angular/platform-browser-dynamic": "~7.1.0",
-    "@angular/router": "~7.1.0",
-    "@angular/http": "~7.1.0",
-    "core-js": "^2.5.4",
+    "@angular/animations": "~7.2.1",
+    "@angular/common": "~7.2.1",
+    "@angular/compiler": "~7.2.1",
+    "@angular/core": "~7.2.1",
+    "@angular/forms": "~7.2.1",
+    "@angular/platform-browser": "~7.2.1",
+    "@angular/platform-browser-dynamic": "~7.2.1",
+    "@angular/router": "~7.2.1",
+    "@angular/http": "~7.2.1",
+    "core-js": "^2.6.2",
     "rxjs": "~6.3.3",
     "tslib": "^1.9.3",
-    "zone.js": "~0.8.26",
+    "zone.js": "~0.8.28",
     "event-source-polyfill": "^1.0.5",
     "font-awesome": "^4.7.0",
-    "luxon": "^1.9.0",
-    "ng2-dragula": "^2.1.1",
-    "dragula": "^3.7.2",
+    "luxon": "^1.10.0",
     "node-sass": "^4.11.0",
     "ts-helpers": "1.1.2",
     "angular2-qrcode": "2.0.1",
@@ -41,10 +39,10 @@
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "^0.12.2",
-    "@angular/cli": "^7.1.4",
-    "@angular/compiler-cli": "^7.1.4",
-    "@types/jasmine": "3.3.5",
-    "@types/node": "~8.9.4",
+    "@angular/cli": "^7.2.2",
+    "@angular/compiler-cli": "^7.2.1",
+    "@types/jasmine": "3.3.7",
+    "@types/node": "~10.12.18",
     "@types/virtual-keyboard": "^1.26.0",
     "codelyzer": "~4.5.0",
     "extract-text-webpack-plugin": "^3.0.2",
@@ -53,11 +51,11 @@
     "karma": "~3.1.1",
     "karma-chrome-launcher": "~2.2.0",
     "karma-coverage-istanbul-reporter": "~2.0.1",
-    "karma-jasmine": "~1.1.2",
-    "karma-jasmine-html-reporter": "^0.2.2",
+    "karma-jasmine": "~2.0.1",
+    "karma-jasmine-html-reporter": "^1.4.0",
     "protractor": "~5.4.0",
     "ts-node": "~7.0.0",
-    "tslint": "~5.11.0",
-    "typescript": "~3.1.6"
+    "tslint": "~5.12.1",
+    "typescript": "~3.2.4"
   }
 }
